pyproject.toml
[project]
name = "default"
version = "0.1.0"
description = "This repository contains configurations to set up a Python development environment using VSCode's Dev Container feature. The environment includes uv and Ruff."

authors = [{ name = "a5chin", email = "a5chin.origin+contact@gmail.com" }]
maintainers = [{ name = "a5chin", email = "a5chin.origin+contact@gmail.com" }]

requires-python = ">=3.9"
readme = "README.md"
license = { file = "LICENSE" }

dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langchain-openai>=0.2.14",
    "langchain>=0.3.13",
    "openai>=1.58.1",
    "pandas>=2.2.3",
    "langchain-google-genai>=2.0.7",
    "langchain-google-vertexai>=2.0.9",
    "python-dotenv>=1.0.1",
    "langsmith>=0.2.6",
    "firecrawl-py>=1.6.8",
    "google-search-results>=2.4.2",
    "googlesearch-python>=1.2.5",
    "langgraph>=0.2.60",
    "retry>=0.9.2",
    "chainlit>=2.0.2",
    "google-cloud-firestore>=2.20.0",
    "google-cloud-storage>=2.19.0",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.0.1",
    "pyright>=1.1.389",
    "pytest>=8.3.3",
    "ruff>=0.8.0",
]
